MGMT-17477: Update dev preview badge when user selects custom OCP release (#2553)
* Update dev preview badge when user select custom OCP release

* Do not store helper text in react/formik state

* Adjusting tests

---------
