+++ libs/ui-lib/lib/ocm/components/clusterConfiguration/OcmOpenShiftVersionSelect.tsx
@@ -1,69 +1,12 @@
 import React from 'react';
-import { ExclamationCircleIcon } from '@patternfly/react-icons/dist/js/icons/exclamation-circle-icon';
-import { ExclamationTriangleIcon } from '@patternfly/react-icons/dist/js/icons/exclamation-triangle-icon';
-import { ExternalLinkAltIcon } from '@patternfly/react-icons/dist/js/icons/external-link-alt-icon';
-import { TFunction } from 'i18next';
 import { useTranslation } from '../../../common/hooks/use-translation-wrapper';
-import {
-  OPENSHIFT_LIFE_CYCLE_DATES_LINK,
-  UiIcon,
-  OpenshiftVersionOptionType,
-  ClusterDetailsValues,
-} from '../../../common';
+import { OpenshiftVersionOptionType, ClusterDetailsValues } from '../../../common';
 import { isInOcm } from '../../../common/api';
 import { OpenShiftVersionDropdown } from '../../../common/components/ui/OpenShiftVersionDropdown';
 import { OpenShiftVersionModal } from './OpenShiftVersionModal';
 import { useFormikContext } from 'formik';
-
-const OpenShiftLifeCycleDatesLink = () => {
-  const { t } = useTranslation();
-  return (
-    <a href={OPENSHIFT_LIFE_CYCLE_DATES_LINK} target="_blank" rel="noopener noreferrer">
-      {t('ai:Learn more')} <ExternalLinkAltIcon />
-    </a>
-  );
-};
-
-const getOpenshiftVersionHelperText = (
-  versions: OpenshiftVersionOptionType[],
-  selectedVersionValue: string | undefined,
-  t: TFunction,
-  isModal?: boolean,
-) => {
-  if (!versions.length && !isModal) {
-    return (
-      <>
-        <UiIcon status="danger" size="sm" icon={<ExclamationCircleIcon />} />
-        &nbsp; {t('ai:No release image is available.')}
-      </>
-    );
-  }
-
-  const selectedVersion = versions.find((version) => version.value === selectedVersionValue);
-  if (!selectedVersionValue || !selectedVersion) {
-    return null;
-  }
-
-  if (
-    selectedVersion.supportLevel !== 'production' &&
-    selectedVersion.supportLevel !== 'maintenance'
-  ) {
-    let messageSelectedVersion = t('ai:Please note that this version is not production-ready.');
-    if (selectedVersion.supportLevel === 'end-of-life') {
-      messageSelectedVersion = t('ai:Please note that this version is not maintained anymore.');
-    }
-    return (
-      <>
-        <UiIcon status="warning" icon={<ExclamationTriangleIcon />} />
-        &nbsp;
-        {messageSelectedVersion}&nbsp;
-        <OpenShiftLifeCycleDatesLink />
-      </>
-    );
-  }
-
-  return null;
-};
+import { getOpenshiftVersionHelperText } from './OpenshiftVersionHelperText';
+import { useOpenshiftVersions } from '../../hooks';
 
 type OcmOpenShiftVersionSelectProps = {
   versions: OpenshiftVersionOptionType[];
@@ -106,22 +49,28 @@ const OcmOpenShiftVersionSelect = ({ versions }: OcmOpenShiftVersionSelectProps)
     }
     return [];
   }, [selectOptions, customOpenshiftSelect]);
+
+  const { versions: allVersions } = useOpenshiftVersions(false);
+
+  const getHelperText = (value: string | undefined, inModal?: boolean) => {
+    return getOpenshiftVersionHelperText(allVersions, value, t, inModal);
+  };
+
   return (
     <>
       <OpenShiftVersionDropdown
         name="openshiftVersion"
         items={selectOptions}
         versions={versions}
-        getHelperText={getOpenshiftVersionHelperText}
+        getHelperText={getHelperText}
         showReleasesLink={isInOcm}
         showOpenshiftVersionModal={showOpenshiftVersionModal}
         customItems={updatedSelectOptions}
       />
       {isOpenshiftVersionModalOpen && (
         <OpenShiftVersionModal
-          isOpen={isOpenshiftVersionModalOpen}
           setOpenshiftVersionModalOpen={setIsOpenshiftVersionModalOpen}
-          getHelperText={getOpenshiftVersionHelperText}
+          getHelperText={getHelperText}
         />
       )}
     </>
